+++ src/style.css
@@ -12,14 +12,16 @@ body {
 }
 
 header {
-  padding-left: 30px;
+  text-align: center;
+  margin-bottom: 15vh;
 }
 
 .container {
   width: 80%;
   margin: auto;
   display: flex;
-  justify-content: space-evenly;
+  justify-content: center;
+  gap: 5rem;
   border: 1px red dashed;
 }
 
@@ -47,7 +49,7 @@ header {
 
 .dynamic-box {
   border: 1px solid black;
-  /* display: none; */
+  display: none;
 }
 
 .dynamic-div:nth-child(even) {
